Oprava chyby: recenzent mohl publikovat clanky

+++ controllers/ReviewController.class.php
@@ -6,20 +6,23 @@ class ReviewController extends DatabaseConnection
 	{
 		parent::__construct();
 
-		if(isset($_GET["remove"]))
+		if(Sessions::isAdmin())
 		{
-			$this->dbArticles->deleteArticle($_GET["remove"]);
-			Sessions::redirect("index.php?web=review");
-		}
-		else if(isset($_GET["publish"]))
-		{
-			$this->dbArticles->publishArticle($_GET["publish"], true);
-			Sessions::redirect("index.php?web=review");
-		}
-		else if(isset($_GET["unpublish"]))
-		{
-			$this->dbArticles->publishArticle($_GET["unpublish"], false);
-			Sessions::redirect("index.php?web=review");
+			if(isset($_GET["remove"]))
+			{
+				$this->dbArticles->deleteArticle($_GET["remove"]);
+				Sessions::redirect("index.php?web=review");
+			}
+			else if(isset($_GET["publish"]))
+			{
+				$this->dbArticles->publishArticle($_GET["publish"], true);
+				Sessions::redirect("index.php?web=review");
+			}
+			else if(isset($_GET["unpublish"]))
+			{
+				$this->dbArticles->publishArticle($_GET["unpublish"], false);
+				Sessions::redirect("index.php?web=review");
+			}
 		}
 
 		$this->data = $data;
